Require to define batch loaders before assigning to attributes
https://github.com/aglushkov/serega/issues/136

This way we can find error about not existing batch loader earlier.

Most of the time user will not reuse loaders, as they are quite unique by nature.
But if they want to reuse it, they will probably define it in parent serializer, so it should not be a problem

Also prefferable way to use batch loaders is to provide callable object directly.